cdk/requirements.txt
attrs==21.4.0
aws-cdk-lib==2.17.0
aws-cdk.aws-batch-alpha==2.17.0a0
boto3==1.21.45
botocore==1.24.45
cattrs==1.10.0
constructs==10.0.92
jmespath==1.0.0
jsii==1.55.1
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.9